🚚 Rename `globalCluster` to fix typo in file name
